refactor[silero_vad]: migrate ReferenceVAD to blocks module; remove redundant tests; add forward and load methods with updated configurations
