+++ .idea/runConfigurations/Test_kits__speech__silero_vad__blocks__reference__tests__test_load_forward_pretrained.xml
@@ -1,6 +1,7 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Test kits::speech::silero_vad::pretrained::tests::test_load_pretrained" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="test --package bunsen --lib --features cuda -- kits::speech::silero_vad::pretrained::tests::test_load_forward_pretrained --ignored --exact" />
+  <configuration default="false" name="Test kits::speech::silero_vad::blocks::reference::tests::test_load_forward_pretrained" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
+    <option name="buildProfileId" value="test" />
+    <option name="command" value="test --package bunsen --lib --features cuda -- kits::speech::silero_vad::blocks::reference::tests::test_load_forward_pretrained --exact --ignored" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs />
     <option name="emulateTerminal" value="true" />
