+++ crates/bunsen/src/kits/speech/silero_vad/blocks/mod.rs
@@ -7,3 +7,6 @@ mod module;
 pub use container::*;
 #[doc(inline)]
 pub use module::*;
+
+#[cfg(any(test, feature = "testing"))]
+pub mod reference;
