+++ crates/bunsen/src/kits/speech/silero_vad/pretrained/mod.rs
@@ -82,97 +82,3 @@ impl<B: Backend> SileroVad16x8<B> {
         Ok(SileroVad16x8 { vad16, vad8 })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    use burn::{
-        Tensor,
-        tensor::{
-            Distribution,
-            Tolerance,
-            backend::BackendTypes,
-        },
-    };
-
-    use super::*;
-    use crate::{
-        errors::*,
-        kits::speech::silero_vad::ReferenceVAD,
-        support::testing::PerformanceBackend,
-    };
-
-    /// A valid chunk length for the given sample rate (standard Silero chunk).
-    pub fn chunk_samples(sample_rate: usize) -> usize {
-        match sample_rate {
-            16000 => 512,
-            8000 => 256,
-            other => panic!("no test chunk for {other}"),
-        }
-    }
-
-    fn silero_burnpack_path() -> PathBuf {
-        PathBuf::from("silero_vad_op18_ifless.bpk")
-    }
-
-    #[test]
-    fn test_load_forward_pretrained() {
-        type B = PerformanceBackend;
-        type F = <B as BackendTypes>::FloatElem;
-
-        let path = silero_burnpack_path();
-        let device = Default::default();
-
-        let s_mod: SileroVad16x8<B> =
-            SileroVad16x8::load_from_burnpack(&path, &device).ok_or_panic();
-
-        let r_mod: ReferenceVAD<B> = ReferenceVAD::from_file(path.to_str().unwrap(), &device);
-
-        let batch = 2;
-        let state = Tensor::zeros([2, batch, 128], &device);
-
-        // 16khz
-        {
-            let sample_rate = 16000;
-            let input = Tensor::<B, 2>::random(
-                [batch, chunk_samples(sample_rate)],
-                Distribution::Default,
-                &device,
-            );
-
-            let (s_out, s_state) = s_mod.forward(input.clone(), sample_rate, state.clone());
-            let (r_out, r_state) = r_mod.forward(input.clone(), sample_rate, state.clone());
-
-            s_out
-                .to_data()
-                .assert_approx_eq::<F>(&r_out.to_data(), Tolerance::default());
-
-            s_state
-                .to_data()
-                .assert_approx_eq::<F>(&r_state.to_data(), Tolerance::default());
-        }
-
-        // 8khz
-        {
-            let sample_rate = 8000;
-
-            let input = Tensor::<B, 2>::random(
-                [batch, chunk_samples(sample_rate)],
-                Distribution::Default,
-                &device,
-            );
-
-            let (s_out, s_state) = s_mod.forward(input.clone(), sample_rate, state.clone());
-            let (r_out, r_state) = r_mod.forward(input.clone(), sample_rate, state.clone());
-
-            s_out
-                .to_data()
-                .assert_approx_eq::<F>(&r_out.to_data(), Tolerance::default());
-
-            s_state
-                .to_data()
-                .assert_approx_eq::<F>(&r_state.to_data(), Tolerance::default());
-        }
-    }
-}
